requirements.txt
Django==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
Faker==4.14.2
mysqlclient==2.0.1